move env.js out of sails-specific config folder
